chore: Remove use of namespace CAD
Replaced CAD namespace with CAD prefix
