+++ Assets/Ctrl + Alt + Defeat/Scripts/CAD_SmartTank.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 684a07a323493d048869d8727365dfc1
+guid: 15869b8e55b4b2548bcfd7719610a2fb
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
