+++ Assets/Ctrl + Alt + Defeat/Scripts/State Machine/Node Editor/CAD_StateMachine.cs
@@ -1,0 +1,8 @@
+using UnityEngine;
+using XNode;
+
+[CreateAssetMenu(menuName = "AI/State Machine")]
+public class CAD_StateMachine : NodeGraph
+{
+    public XNode.Node CurrentNode { get; set; }
+}
