+++ Assets/Ctrl + Alt + Defeat/Scripts/State Machine/Node Editor/Editor/CAD_StateMachineGraphEditor.cs.meta
