+++ Assets/Ctrl + Alt + Defeat/Scripts/State Machine/Node Editor/Nodes/CAD_EntryNode.cs
@@ -1,0 +1,9 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+[DisallowMultipleNodes, NodeTint("#00ff00")]
+public class CAD_EntryNode : XNode.Node
+{
+    [Output] private int m_OutputNode;
+}
