+++ Assets/Ctrl + Alt + Defeat/Scripts/State Machine/Node Editor/Nodes/CAD_ExitNode.cs
@@ -1,0 +1,9 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+[DisallowMultipleNodes, NodeTint("#ff0000")]
+public class CAD_ExitNode : XNode.Node
+{
+    [Input] private int m_InputNode;
+}
