+++ Assets/Ctrl + Alt + Defeat/Scripts/State Machine/Node Editor/Nodes/CAD_ExitNode.cs.meta
