+++ Assets/Ctrl + Alt + Defeat/Scripts/State Machine/Node Editor/Nodes/Editor/CAD_EntryNodeEditor.cs
@@ -1,0 +1,17 @@
+using UnityEngine;
+using XNodeEditor;
+
+
+[CustomNodeEditor(typeof(CAD_EntryNode))]
+public class CAD_EntryNodeEditor : NodeEditor
+{
+    public override void OnBodyGUI()
+    {
+        var node = target as CAD_EntryNode;
+
+        NodeEditorGUILayout.PortField(new GUIContent(" "), node.GetOutputPort("m_OutputNode"), GUILayout.MinWidth(0));
+
+        base.OnBodyGUI();
+    }
+}
+
