+++ Assets/Ctrl + Alt + Defeat/Scripts/State Machine/Node Editor/Nodes/Editor/CAD_EntryNodeEditor.cs.meta
