--- Assets/Ctrl + Alt + Defeat/Scripts/State Machine/Node Editor/Nodes/Editor/EntryNodeEditor.cs
@@ -1,18 +1,0 @@
-using UnityEngine;
-using XNodeEditor;
-
-namespace CAD
-{
-    [CustomNodeEditor(typeof(EntryNode))]
-    public class EntryNodeEditor : NodeEditor
-    {
-        public override void OnBodyGUI()
-        {
-            var node = target as EntryNode;
-
-            NodeEditorGUILayout.PortField(new GUIContent(" "), node.GetOutputPort("m_OutputNode"), GUILayout.MinWidth(0));
-
-            base.OnBodyGUI();
-        }
-    }
-}
