--- Assets/Ctrl + Alt + Defeat/Scripts/State Machine/Node Editor/Nodes/EntryNode.cs
@@ -1,12 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-namespace CAD
-{
-    [DisallowMultipleNodes, NodeTint("#00ff00")]
-    public class EntryNode : XNode.Node
-    {
-        [Output] private int m_OutputNode;
-    }
-}
