--- Assets/Ctrl + Alt + Defeat/Scripts/State Machine/Node Editor/Nodes/ExitNode.cs
@@ -1,12 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-namespace CAD
-{
-    [DisallowMultipleNodes, NodeTint("#ff0000")]
-    public class ExitNode : XNode.Node
-    {
-        [Input] private int m_InputNode;
-    }
-}
