--- Assets/Ctrl + Alt + Defeat/Scripts/State Machine/Node Editor/StateMachine.cs
@@ -1,11 +1,0 @@
-using UnityEngine;
-using XNode;
-
-namespace CAD
-{
-    [CreateAssetMenu(menuName = "AI/State Machine")]
-    public class StateMachine : NodeGraph
-    {
-        public XNode.Node CurrentNode { get; set; }
-    }
-}
